sum up fakes in fakes file and add fake for validationResult
