--- src/xcode/ENA/HealthCertificateToolkit/Sources/HealthCertificateToolkit/Validation/ValidationRulesCBORDataFake.swift
@@ -1,17 +1,0 @@
-//
-// 🦠 Corona-Warn-App
-//
-
-import Foundation
-import CertLogic
-import SwiftCBOR
-
-func rulesCBORDataFake() throws -> Data {
-    let rules = [
-        Rule.fake(),
-        Rule.fake(),
-        Rule.fake()
-    ]
-
-    return try CodableCBOREncoder().encode(rules)
-}
